+++ src/app/record/page.tsx
@@ -2,7 +2,7 @@ import { NavBar } from "@/components/common/NavBar";
 import { Header } from "@/components/dailyRecord/Header";
 import { RecordMissionCard } from "@/components/dailyRecord/MissionCard";
 import { RecordMissionFeed } from "@/components/dailyRecord/MissionFeed";
-import { SharedDiaryCard } from "@/components/dailyRecord/ShareDiaryCard";
+import { SharedDiaryCard } from "@/components/dailyRecord/SharedDiaryCard";
 
 const DailyRecordPage = () => {
   return (
