+++ src/types/diary.type.ts
@@ -1,0 +1,12 @@
+import { Emotion } from "./emotion.type";
+
+export type Diary = {
+  id: number;
+  userId: number;
+  userName: string;
+  profile: string;
+  emotion: Emotion;
+  text: string;
+  image: string;
+  createdAt: string;
+};
